:recycle: refactor: API response type 정의 [#32]
